Moving expense group specific components
