+++ src/components/Expense_Groups/Expense_Group_Child_Edit_View.js
@@ -1,5 +1,5 @@
 // Helper Functions
-import { capitalizeFirstLetter } from "../helpers/helpers";
+import { capitalizeFirstLetter } from "../../helpers/helpers";
 
 
 // Material UI Components
